# docs/swagger.yaml
basePath: /
definitions:
  book.Book:
    description: Модель книги для системы обмена книгами
    properties:
      author:
        description: |-
          @Description Автор книги
          @example F. Scott Fitzgerald
        type: string
      created_at:
        description: |-
          @Description Время создания записи
          @example 2025-04-28T12:00:00Z
        type: string
      description:
        description: |-
          @Description Описание книги
          @example A story of the fabulously wealthy Jay Gatsby and his love for the beautiful Daisy Buchanan.
        type: string
      id:
        description: |-
          @Description Уникальный идентификатор
          @example 1
        type: integer
      photos:
        description: |-
          @Description Фотографии книги в формате base64
          @example ["data:image/jpeg;base64,/9j/4AAQSkZJRg..."]
        type: string
      price:
        description: |-
          @Description Цена книги
          @example 19.99
        type: number
      state:
        allOf:
        - $ref: '#/definitions/state.State'
        description: '@Description Состояние книги'
      state_id:
        description: |-
          @Description ID состояния книги
          @example 1
        type: integer
      tags:
        description: '@Description Теги книги'
        items:
          $ref: '#/definitions/tag.Tag'
        type: array
      title:
        description: |-
          @Description Название книги
          @example The Great Gatsby
        type: string
      updated_at:
        description: |-
          @Description Время последнего обновления записи
          @example 2025-04-28T12:00:00Z
        type: string
      user:
        allOf:
        - $ref: '#/definitions/user.User'
        description: '@Description Пользователь-владелец'
      user_id:
        description: |-
          @Description ID пользователя-владельца
          @example 1
        type: integer
    type: object
...
  state.State:
    description: Модель состояния книги
    properties:
      created_at:
        description: |-
          @Description Время создания записи
          @example 2025-04-28T12:00:00Z
        type: string
      id:
        description: |-
          @Description Уникальный идентификатор
          @example 1
        type: integer
      name:
        description: |-
          @Description Название состояния
          @example available
        type: string
      updated_at:
        description: |-
          @Description Время последнего обновления записи
          @example 2025-04-28T12:00:00Z
        type: string
    type: object
...
  tag.Tag:
    description: Модель тега для категоризации книг
    properties:
      created_at:
        description: |-
          @Description Время создания записи
          @example 2025-04-28T12:00:00Z
        type: string
      id:
        description: |-
          @Description Уникальный идентификатор
          @example 1
        type: integer
      name:
        description: |-
          @Description Название тега
          @example fiction
        type: string
      updated_at:
        description: |-
          @Description Время последнего обновления записи
          @example 2025-04-28T12:00:00Z
        type: string
    type: object
  user.CreateUserDTO:
    properties:
      login:
        maxLength: 50
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - login
    - password
    - username
    type: object
  user.LoginDTO:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  user.TokenResponse:
    properties:
      token:
        type: string
    type: object
  user.UpdateUserDTO:
    properties:
      avatar:
        type: string
      description:
        maxLength: 1000
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  user.User:
    description: Модель пользователя системы обмена книгами
    properties:
      avatar:
        description: |-
          @Description Аватар пользователя в формате base64
          @example data:image/jpeg;base64,/9j/4AAQSkZJRg...
        type: string
      books:
        description: |-
          @Description Книги пользователя
          @example [{"id": 1, "title": "The Great Gatsby"}]
        items:
          $ref: '#/definitions/book.Book'
        type: array
      created_at:
        description: |-
          @Description Время создания аккаунта
          @example 2025-04-28T12:00:00Z
        type: string
      description:
        description: |-
          @Description Описание пользователя
          @example Book lover and collector
        type: string
      id:
        description: |-
          @Description Уникальный идентификатор
          @example 1
        type: integer
      login:
        description: |-
          @Description Логин пользователя
          @example john_doe
        type: string
      updated_at:
        description: |-
          @Description Время последнего обновления аккаунта
          @example 2025-04-28T12:00:00Z
        type: string
      username:
        description: |-
          @Description Имя пользователя
          @example john_doe
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API for book trading system with tag support
  title: Book Trading API
  version: "1.0"
paths:
...
  /api/v1/users/{id}:
    delete:
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
...
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: Login user and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.TokenResponse'
      summary: Login user
      tags:
      - users
  /api/v1/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
      summary: Register new user
      tags:
      - users
swagger: "2.0"